% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alarmfilter.R
\name{alarmfilter}
\alias{alarmfilter}
\title{Filter alarm signals to reduce errors from transient events}
\usage{
alarmfilter(recovalarm.obj, dmin.dist, dmin.recov = dmin.dist)
}
\arguments{
\item{recovalarm.obj}{output from \code{recovalarm}}

\item{dmin.dist}{a minimum duration for disturbances, in days}

\item{dmin.recov}{a minimum duration for recovery events, in days}
}
\value{
\code{alarmfilter} returns a data frame containing the columns:
\item{dist.date} time a disturbance was detected
\item{recov.date} time a recovery was detected
\item{tdiff} difference between \code{recov} and \code{disturb}; the recovery time.
This is the same format as output from \code{recovalarm}, but there will be fewer rows.
}
\description{
\code{alarmfilter} filters output from \code{recovalarm}
}
\details{
\code{alarmfilter} merges disturbances separated by recoveries shorter than \code{dmin.recov}
and then removes disturbances shorter than \code{dmin.dist}
}
