% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimal_doy.R
\name{decimal_doy}
\alias{decimal_doy}
\title{Calculate the decimal day of year}
\usage{
decimal_doy(datetime)
}
\arguments{
\item{datetime}{a date-time object in POSIX format}
}
\value{
\code{decimal_doy} returns a numeric vector with the decimal day of year
}
\description{
The decimal day of year uses hours, minutes, and seconds to compute the fractional
day, expressed as a decimal.
}
\details{
\code{decimal_doy} computes the decimal day of year
}
\author{
Jonathan Walter, \email{jaw3es@virginia.edu}
}
