% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwdistdiffzscore.R
\name{mwdistdiffz}
\alias{mwdistdiffz}
\title{Calculate differences in continuous distribution functions between reference and moving window observations}
\usage{
mwdistdiffz(
  testy,
  refy,
  wwidth,
  refwidth = NULL,
  dx = 0.01,
  stride = 1,
  dmin = 0.5
)
}
\arguments{
\item{testy}{a data frame representing the "test" period, i.e., the period over which to search for disturbance and recovery.
It must contain the columns \code{tt} and \code{yy}.
\code{tt} contains numeric times assumed to be in units of days, or POSIXct (date) format corresponding to observations of \code{yy}.
If \code{is.numeric(tt)}, tt is taken to be in units of day-of-year, with January 1 = doy 1.
\code{yy} is a numeric time series.}

\item{refy}{a data frame representiong the refrence period that \code{testy} is compared to.
It must contain the columns \code{tt} and \code{yy}, as in \code{testy}.}

\item{wwidth}{the moving window width, in number of time steps}

\item{refwidth}{the width of the rolling reference window, in number of time steps; if \code{NULL} do not use rolling reference window}

\item{dx}{increment between values at which to evaluate differences between the cdf for \code{yy} and \code{refy}}

\item{stride}{number of time steps by which the moving window advances}

\item{dmin}{the fraction of data that must be present (i.e., non-NA) in test (and, if applicable, adaptive reference) moving windows to procede with computations.}
}
\value{
\code{mwdistdiff} returns a data frame containing the columns:
\item{tleft}{the time corresponding to the beginning (left edge) of the moving window}
\item{tright}{the time corresponding to the end (right edge) of the moving window}
\item{ddiff}{differences from the reference distritbution}
\item{zz}{Z-scores representing the strength of excursion from reference}
}
\description{
\code{mwdistdiffz_ks} computes differences between the continuous distribution functions (cdf) for observations within
a moving window and a reference distribution, and the z-scores of differences relative to samples of the reference distribution.
It is used for identifying recovery times from disturbance in time series data.
}
\details{
The value suppplied to \code{refwidth} determines whether a rolling reference window will be used,
or if all of refy will be used as the reference period.
\code{testy} and \code{refy} must be regularly spaced, ordered time series. If there are gaps in the time series, fill
them with NAs prior to applying this function. If there are insufficient data then output will contain \code{NA}s.
\code{dmin} takes into account NAs in time series as well as reference windows overlapping the edges or gaps in the time series.
Currently, the year is assumed to be 365 days long (i.e., there is no internal handling of leap years).
}
\examples{
#need to add some

}
\author{
Jonathan Walter, \email{jaw3es@virginia.edu}
}
