% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mwddz.R
\name{plot_mwddz}
\alias{plot_mwddz}
\title{Plot output (& optinonally input) from mwdistdiffz()}
\usage{
plot_mwddz(
  mwddz,
  window_side = "right",
  diff_ts = "ddiff",
  testy = NULL,
  refy = NULL
)
}
\arguments{
\item{mwddz}{data frame, output from call to mwdistdiffz()}

\item{window_side}{options are "right"  (default), "left", or "center"; which time index of each window should be used}

\item{diff_ts}{options are "ddiff" or "zz"; should the actual difference between distributions be plotted ("ddiff"), or the z-score ("zz")}

\item{testy}{test data frame with columns 'tt' and 'yy' that went into mwdistdiffz()}

\item{refy}{reference data frame with columns 'tt' and 'yy' that went into mwdistdiffz()}
}
\value{
a plot of the difference between test and reference distribution through time
}
\description{
Plot output (& optinonally input) from mwdistdiffz()
}
